requirements.txt
Django==1.11
django-multitenant==1.1.0
django-tenant-schemas==1.9.0
djangorestframework==3.8.2
psycopg2==2.7.6.1
pytz==2018.9
